+++ test/invariants/PYUSDXInvariants.t.sol
@@ -30,7 +30,7 @@ contract PYUSDXInvariants is PYUSDXBaseUnitTest {
         testAccounts[3] = david;
 
         for (uint256 i = 0; i < testAccounts.length; i++) {
-            (uint32 earnerRate_, , ) = pyusdx.getAccountEarningInfo(testAccounts[i]);
+            (uint16 earnerRate_, , ) = pyusdx.getAccountEarningInfo(testAccounts[i]);
             if (earnerRate_ == 0) {
                 assertEq(pyusdx.earningPrincipalOf(testAccounts[i]), 0, "Non-earning account has non-zero principal");
             }
@@ -61,7 +61,7 @@ contract PYUSDXInvariants is PYUSDXBaseUnitTest {
         testAccounts[3] = david;
 
         for (uint256 i = 0; i < testAccounts.length; i++) {
-            (uint32 earnerRate_, , ) = pyusdx.getAccountEarningInfo(testAccounts[i]);
+            (uint16 earnerRate_, , ) = pyusdx.getAccountEarningInfo(testAccounts[i]);
             if (earnerRate_ > 0) {
                 uint256 balance = pyusdx.balanceOf(testAccounts[i]);
                 uint112 principal = pyusdx.earningPrincipalOf(testAccounts[i]);
